Feat(Adapter): Add session adapter
